Added.
* Check current token.

How It works?
Firstly, the library checks if "token.json" exists on the hard drive.
Secondly, if you have this file but "access_token" is unalive the library will use "refresh_token" to get a new "access_token".

You need to check token file before authentication manually.

* HttpReq (implements the interface)

It a class which you can assist to make a async request in one line.